task2/requirements.txt
requests==2.26.0
boto3==1.15.3
Pillow==8.4.0